# docs/openapi.yaml
openapi: 3.1.0
info:
  title: OneAce API
  version: 1.0.0
  description: |
    OneAce is a comprehensive inventory management and stock counting platform.

    ## Authentication
    Most endpoints require authentication via session cookie (from Better Auth).
    Some endpoints (webhooks, cron, health) are public but have additional security measures.

    ## Rate Limiting
    Rate limits vary by endpoint. When rate-limited, the response includes:
    - `Retry-After` header: seconds until the next request is allowed
    - `X-RateLimit-*` headers: current limit status

    ## Error Responses
    All error responses include an `error` field describing the issue.
    Common status codes:
    - 400: Invalid request
    - 401: Unauthenticated or invalid credentials
    - 403: Forbidden / insufficient permissions
    - 404: Resource not found
    - 429: Rate limited
    - 500: Internal server error
    - 503: Service unavailable
servers:
  - url: https://oneace-next-local.vercel.app/api
    description: Production
  - url: http://localhost:3000/api
    description: Local development
...
paths:
  /account/delete:
    post:
      tags:
        - Account
      summary: Delete user account
      description: |
        Permanently delete the authenticated user account. Requires confirmation phrase.
        Requires active session cookie. Rate limited: 1 per hour per user.
        User cannot be owner of any organization.
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                confirmation:
                  type: string
                  description: Must be exactly "DELETE MY ACCOUNT"
                  example: "DELETE MY ACCOUNT"
              required:
                - confirmation
      responses:
        200:
          description: Account successfully deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  message:
                    type: string
        400:
          description: Invalid confirmation phrase
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        403:
          description: User is owner of an organization
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  ownedOrganizations:
                    type: array
                    items:
                      type: string
        429:
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        401:
          description: Unauthenticated
        500:
          description: Internal server error
...
  /billing/portal:
    post:
      tags:
        - Billing
      summary: Create billing portal session
      description: |
        Create a Billing Portal session for subscription management.
        Requires OWNER or ADMIN role and existing Stripe customer.
      security:
        - sessionCookie: []
      responses:
        200:
          description: Portal session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    description: Billing portal URL
        401:
          description: Unauthenticated
        403:
          description: Insufficient permissions
        404:
          description: No billing account found
        503:
          description: Stripe not configured
...
  /cron/stock-count-triggers:
    get:
      tags:
        - Cron
      summary: Trigger stock count pass
      description: |
        Executes stock count trigger pass for all recurring templates.
        Called every 15 minutes by Vercel Cron or external scheduler.
        Requires Bearer token authorization.
      security:
        - bearerToken: []
      responses:
        200:
          description: Trigger pass completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  triggered:
                    type: integer
                    description: Count of templates triggered
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        organizationId:
                          type: string
                        templateId:
                          type: string
                        countId:
                          type: string
                          nullable: true
                        reason:
                          type: string
        401:
          description: Unauthorized (invalid token)
        503:
          description: CRON_SECRET not configured
...
  /integrations/shopify/callback:
    get:
      tags:
        - Integrations
      summary: Shopify OAuth callback
      description: |
        Receives OAuth callback from Shopify.
        Exchanges authorization code for access token and stores integration.
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
        - name: shop
          in: query
          required: true
          schema:
            type: string
          description: Shopify store domain
        - name: hmac
          in: query
          required: true
          schema:
            type: string
      responses:
        302:
          description: Redirect to settings page with success/error status
        400:
          description: Missing OAuth parameters
        403:
          description: Invalid state signature
  /integrations/shopify/webhooks:
    post:
      tags:
        - Integrations
      summary: Shopify webhook receiver
      description: |
        Receives topic notifications from Shopify.
        Public endpoint. Requires HMAC-SHA256 signature verification.
      parameters:
        - name: x-shopify-hmac-sha256
          in: header
          required: true
          schema:
            type: string
        - name: x-shopify-topic
          in: header
          required: true
          schema:
            type: string
        - name: x-shopify-shop-domain
          in: header
          required: true
          schema:
            type: string
      responses:
        200:
          description: Webhook processed
        400:
          description: Missing headers or invalid JSON
        401:
          description: Invalid signature
        503:
          description: Webhook not configured
...
components:
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: Better Auth session cookie
    bearerToken:
      type: http
      scheme: bearer
      description: Bearer token for cron endpoints (CRON_SECRET)